requirements.txt
asgiref==3.8.1
bip32utils==0.3.post4
boto3==1.35.57
botocore==1.35.57
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.6.0
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==4.2.14
django-allauth==0.63.6
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.0.0
ecdsa==0.19.0
Faker==26.0.0
idna==3.7
jmespath==1.0.1
mailjet-rest==1.3.4
mininet==2.3.0.dev6
ns==0.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20
whitenoise==6.7.0
gunicorn==20.1.0
